+++ eventmetrics/parsers/durationCalculations_test.go
@@ -3,12 +3,16 @@ package parsers
 import (
 	"errors"
 	"fmt"
+	"io/ioutil"
+	"log"
 	"testing"
 	"time"
 
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/xmidt-org/interpreter"
+	"github.com/xmidt-org/touchstone"
 	"go.uber.org/zap"
 )
 
@@ -252,4 +256,102 @@ func TestEventToCurrentCalculator(t *testing.T) {
 			assert.Equal(tc.expectedErr, err)
 		})
 	}
+}
+
+func TestCreateDurationCalculators(t *testing.T) {
+	tests := []struct {
+		description string
+		configs     []TimeElapsedConfig
+		expectedErr error
+	}{
+		{
+			description: "success",
+			configs: []TimeElapsedConfig{
+				TimeElapsedConfig{
+					Name:        "test",
+					SessionType: "current",
+					EventType:   "test-event-type",
+				},
+				TimeElapsedConfig{
+					Name:        "test1",
+					SessionType: "previous",
+					EventType:   "test-event-type2",
+				},
+			},
+		},
+		{
+			description: "duplicate Name",
+			configs: []TimeElapsedConfig{
+				TimeElapsedConfig{
+					Name:        "test",
+					SessionType: "current",
+					EventType:   "test-event-type",
+				},
+				TimeElapsedConfig{
+					Name:        "test",
+					SessionType: "previous",
+					EventType:   "test-event-type2",
+				},
+			},
+			expectedErr: errNewHistogram,
+		},
+		{
+			description: "blank name",
+			configs: []TimeElapsedConfig{
+				TimeElapsedConfig{
+					Name:        "test",
+					SessionType: "current",
+					EventType:   "test-event-type",
+				},
+				TimeElapsedConfig{
+					Name:        "",
+					SessionType: "previous",
+					EventType:   "test-event-type2",
+				},
+			},
+			expectedErr: errBlankHistogramName,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.description, func(t *testing.T) {
+			assert := assert.New(t)
+			testFactory := touchstone.NewFactory(touchstone.Config{}, log.New(ioutil.Discard, "", 0), prometheus.NewPedanticRegistry())
+
+			testMeasures := Measures{TimeElapsedHistograms: make(map[string]prometheus.ObserverVec)}
+			durationCalculators, err := createDurationCalculators(testFactory, tc.configs, testMeasures, RebootLoggerIn{Logger: zap.NewNop()})
+
+			if tc.expectedErr != nil {
+				assert.True(errors.Is(err, tc.expectedErr))
+			} else {
+				assert.NotNil(durationCalculators)
+				assert.Equal(len(tc.configs), len(durationCalculators))
+				for _, config := range tc.configs {
+					histogram, found := testMeasures.TimeElapsedHistograms[config.Name]
+					assert.True(found)
+					assert.NotNil(histogram)
+				}
+			}
+		})
+	}
+}
+
+func TestCreateDurationCalculatorsHistogramErr(t *testing.T) {
+	assert := assert.New(t)
+	testFactory := touchstone.NewFactory(touchstone.Config{}, log.New(ioutil.Discard, "", 0), prometheus.NewPedanticRegistry())
+	testMeasures := Measures{TimeElapsedHistograms: make(map[string]prometheus.ObserverVec)}
+	testName := "test_hist"
+	config := TimeElapsedConfig{
+		Name: testName,
+	}
+	options := prometheus.HistogramOpts{
+		Name:    testName,
+		Help:    "test_help",
+		Buckets: []float64{60, 120, 180, 240, 300, 360, 420, 480, 540, 600, 900, 1200, 1500, 1800, 3600, 7200, 14400, 21600},
+	}
+
+	testMeasures.addTimeElapsedHistogram(testFactory, options)
+	durationCalculators, err := createDurationCalculators(testFactory, []TimeElapsedConfig{config}, testMeasures, RebootLoggerIn{Logger: zap.NewNop()})
+	assert.True(errors.Is(err, errNewHistogram))
+	assert.Nil(durationCalculators)
 }
